Move ResolverMethod into Method
